+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nieuw guard spawning systeem toegevoegd om makkelijker en beter enemy NPCs te spawnen
 - Nieuwe criminele drugs activiteit toegevoegd
 - Commands toegevoegd om gangs te creeren/verwijderen
+- Verschil tussen twitter reply en retweet toegevoegd
 
 ### Changed
 - Override CID entry toegevoegd om kofi vehicles te kunnen assignen aan offline vehicles